Poprawa obliczania wartości funkcji dla TSP + wymiana list w reprezentacji fenotypu dla Problemów Optymalizacyjnych

+++ AlgorytmyDoTTP/AlgorytmyDoTTP/Struktura/ProblemyOptymalizacyjne/Abstrakcyjny/IPomocniczy.cs
@@ -6,31 +6,31 @@
         /// Metoda zwraca wagę przedmiotu
         /// </summary>
         /// <returns>Waga przedmiotu</returns>
-        double ZwrocWage();
+        float ZwrocWage();
 
         /// <summary>
         /// Metoda zwraca wartość przedmiotu
         /// </summary>
         /// <returns>Wartość przedmiotu</returns>
-        double ZwrocWartosc();
+        float ZwrocWartosc();
 
         /// <summary>
         /// Metoda zwraca miasto startowe na odcinku
         /// </summary>
         /// <returns>Numer miasta</returns>
-        short ZwrocOd();
+        ushort ZwrocOd();
 
         /// <summary>
         /// Metoda zwraca miasto końcowe na odcinku
         /// </summary>
         /// <returns>Numer miasta</returns>
-        short ZwrocDo();
+        ushort ZwrocDo();
 
         /// <summary>
         /// Metoda zwraca długość odcinka
         /// </summary>
         /// <returns>Długość odcinka</returns>
-        double ZwrocDlugosc();
+        float ZwrocDlugosc();
 
         /// <summary>
         /// Metoda zwraca dostępne 
@@ -38,6 +38,6 @@
         /// <param name="dostepnePrzedmioty">Dostępne przedmioty w danym mieście w notacji tekstowej</param>
         /// <param name="iloscPrzedmiotow">Ilość dostępnych przedmiotów we wszystkich miastach</param>
         /// <returns>Dostępne przedmioty w notacji tablicowej</returns>
-        ushort[] ZwrocPrzedmioty(string dostepnePrzedmioty, int iloscPrzedmiotow);
+        ushort[] ZwrocPrzedmioty(string dostepnePrzedmioty, ushort iloscPrzedmiotow);
     }
 }
